scripts/requirements.txt
numpy
jupyter
pandas
psycopg2-binary
factom-api
pegnet-py

